feat: refactoring raw data
