+++ README.md
@@ -2,6 +2,7 @@
 This is a learning repository about DVC Data Version Control and Luigi Pipelines
 
 - luigi, dvc, pre-commit
-- setup https://pre-commit.com/
+- https://luigi.readthedocs.io/
+- setup https://pre-commit.com/, https://pre-commit.com/hooks.html
 - setup https://github.com/Kaggle/kaggle-api
 - `kaggle competitions download -c sentiment-analysis-on-movie-reviews -p data`
