+++ requirements.txt
@@ -2,3 +2,4 @@ pre-commit==3.5.0
 kaggle==1.5.16
 dvc==3.28.0
 luigi==3.4.0
+scikit-learn==1.3.2
